Shrink Token from 24 to 20 bytes: u32 start, derive trivial end offsets
Two layout changes:

- ElementStart, Attribute, Text, Comment and Cdata no longer store an
  explicit end offset. Their end is fully determined by the last
  sub-span (local name, value plus closing quote, text, text plus
  terminator). The new Token::range() derives the document range for
  every variant, and the integration tests now validate all spans
  through it. To keep ends derivable for empty values and CDATA/comment
  bodies, detach()/detach_small() now preserve real offsets of empty
  spans and only map placeholder ("".into()) spans to 0..0.

- `start` is stored as u32 instead of usize, which limits documents to
  4 GiB; longer input now fails with the new
  StreamError::DocumentTooLarge instead of parsing. The per-token checks
  are folded into one branch: if s.pos() fits u32, the start and every
  sub-span offset fit their storage.

size_of::<Token>() drops from 24 to 20 bytes (upstream: 112), so a
collected Vec<Token> shrinks by another 17%. Tokenize-only throughput
pays 1-3% on element-heavy fixtures for the extra checks but remains
20-30% ahead of upstream (dense 0.81, spaced 0.77, dtd 0.79 ratios).

+++ tests/integration/limits.rs
@@ -23,7 +23,7 @@ fn attribute_under_64k_roundtrips() {
             _ => None,
         })
         .unwrap();
-    assert_eq!(attr.1.as_str(&xml, attr.0), value);
+    assert_eq!(attr.1.as_str(&xml, attr.0 as usize), value);
 }
 
 #[test]
@@ -49,10 +49,7 @@ fn pi_content_over_64k_is_an_error() {
 
 #[test]
 fn entity_value_over_64k_is_an_error() {
-    let xml = format!(
-        "<!DOCTYPE d [<!ENTITY e \"{}\">]><r/>",
-        "x".repeat(70_000)
-    );
+    let xml = format!("<!DOCTYPE d [<!ENTITY e \"{}\">]><r/>", "x".repeat(70_000));
     let has_err = tokens(&xml).into_iter().any(|t| t.is_err());
     assert!(has_err);
 }
@@ -69,7 +66,7 @@ fn text_over_64k_roundtrips() {
             _ => None,
         })
         .unwrap();
-    assert_eq!(text.1.as_str(&xml, text.0), body);
+    assert_eq!(text.1.as_str(&xml, text.0 as usize), body);
 }
 
 #[test]
@@ -83,7 +80,7 @@ fn comment_over_64k_roundtrips() {
             _ => None,
         })
         .unwrap();
-    assert_eq!(comment.1.as_str(&xml, comment.0), body);
+    assert_eq!(comment.1.as_str(&xml, comment.0 as usize), body);
 }
 
 #[test]
@@ -97,5 +94,5 @@ fn cdata_over_64k_roundtrips() {
             _ => None,
         })
         .unwrap();
-    assert_eq!(cdata.1.as_str(&xml, cdata.0), body);
+    assert_eq!(cdata.1.as_str(&xml, cdata.0 as usize), body);
 }
